requirements.txt
python-binance==1.0.29
pandas==2.3.3
loguru==0.7.3
websockets==15.0.1
python-dotenv==1.1.1
PyYAML==6.0.1
